✨ feat(weathercard): adds the date and time of the API
Currently, the screen displays "You are in [your city], and it is [time] of [morning/night/evening]", however, the API data section does not show when the displayed data was last updated. It is expected to be updated at the exact moment the user is viewing it, but the API may occasionally introduce some latency.

resolve #12